_archive/invoice-scanner/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "invoice-scanner"
version = "0.1.0"
description = "AI agent that extracts invoice data for Stocksure"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "httpx>=0.26.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
invoice_scanner = ["extraction_prompt.txt"]
